refactor(revision): migrate to /v0/revisions
