+++ pol/api/v0/__init__.py
@@ -1,10 +1,11 @@
 from fastapi import APIRouter
 
-from . import me, person, episode, subject, character
+from . import me, person, episode, subject, revision, character
 
 router = APIRouter()
 router.include_router(subject.router)
 router.include_router(episode.router)
 router.include_router(character.router)
 router.include_router(person.router)
+router.include_router(revision.router)
 router.include_router(me.router)
