+++ pol/curd/exceptions.py
@@ -1,2 +1,6 @@
+from typing import Any
+
+
 class NotFoundError(Exception):
-    pass
+    def __init__(self, details: Any = None):
+        self.details = details
